feat: republish major tag on new npm version (#20)
| 🚥 Fixes #9 |
| :---------------- |

## 🧰 Changes

GitHub Actions allows you to define a tag for an action to use that
doesn't have to be a fully qualified semver version. This means we can
publish v2.3.0 onto npm and then re-push that to v2 as a git tag. This
means we dont have to prompt users to upgrade any time we push a
non-breaking change. This is a pretty common pattern that GitHub does
themselves, and we also do it in rdme too!

So then instead of telling people to do this:

```yaml
- name: GitHub Action
  uses: readmeio/readme-micro@v2.3.0
```

We can just do this
```yaml
- name: GitHub Action
  uses: readmeio/readme-micro@v2
```

## 🧬 QA & Testing

If you run this:

```sh
./bin/set-major-version-tag.js
$ git tag v2 -f -m 'Top-level tag pointing to 2.3.0'
Updated tag 'v2' (was fc99ec6)
```

You'll see how it works. We've automatically wired this up to the
[postversion](https://docs.npmjs.com/cli/v7/commands/npm-version#description)
event which is after the new npm version has been created and the new
git tag is present.